add headers to properties
Allow user to set the headers for the request and set a default
value if undefined. The headers property is an object which must
be in json format like:

```
{"Content-type":"application/json", "Accept": "application/json"}
```

Since dcache has included a way to suppress WWW-Authenticate header,
this patch added "Suppress-WWW-Authenticate": "Suppress" to the list
of default headers.

Adjust the demo page to fit all the new changes.

Also, update to the latest MultipleAjaxRequests.